chore: fix file generator tests (#782)
